#393 移动redis代码到独立模块
