+++ modules/redis/install
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+echo "[INFO] Install jar to local repository."
+
+mvn clean install
