+++ modules/redis/install.bat
@@ -1,0 +1,6 @@
+@echo off
+echo [INFO] Install jar to local repository.
+
+cd %~dp0
+call mvn clean install
+pause
