+++ modules/redis/src/main/java/org/springside/modules/nosql/redis/JedisShardedTemplate.java
