+++ modules/redis/src/main/java/org/springside/modules/nosql/redis/elector/MasterElector.java
