+++ modules/redis/src/main/java/org/springside/modules/nosql/redis/pool/ConnectionInfo.java
