+++ modules/redis/src/main/resources/redis/batchpop.lua
