+++ modules/redis/src/test/java/org/springside/modules/nosql/redis/JedisTemplateTest.java
